Updated to fix Explorer related security issues
The lines added pre-create the directories needed for explorer and allow for reading of the dirs for STAT purposes.  This eliminates the ICH408I messages issues while starting Explorer jars.  This fix pre-creates the following dirs in explorer

In `./explorer-server/wlp/usr/servers/Atlas`
`./workarea/org.eclipse.osgi/169/0/.cp` are created and set with a+rx permissions to allow STATing